detection-fissures/requirements.txt
fastapi==0.75.0
uvicorn==0.18.3
pillow==9.5.0
numpy==1.24.3
python-multipart==0.0.9
